Hotfix for transparency issues.
There was an bug with BetterDiscord to the point where the background image was not transparent to the chat bar and server members list. So I added a piece of code and it fixed the problem. For things are coming to this theme script to make the transparency look even better so stay tuned.